CasasDatabaseService: Fix Local Date Calculation
Use reader.GetFieldValue<DateTimeOffset>(0) instead of reader.GetDateTime(0) to grabe a correct DateTimeOffset with UTC offset. The GetDateTime function will change the "Kind" field to local (not sure whether it is local to the system or local to DB server), and as a result, the value written into the event.csv file are not correct.