Fix A Bug where exception is thrown when change to another date
The bug is introduced by the binding of total events on dataset event page to the count property of EventsInView property.
As EventsInView is modified in async thread during data loading, there is an access conflict between UI thread and the async Task.
As a result, a system Exception is thrown and no further information is given (as the exception is thrown in async Task, the UI thread never receives the exception but the Task exits without loading new data).

The fix is to add a property to DatasetEventsViewModel and update it when filter is updated or new data is loaded (with a dispatcher to run on UI thread).